+++ src/cuda/stream.c
@@ -1,7 +1,6 @@
 #include "include/libcuda_hook.h"
 
 CUresult cuStreamCreate(CUstream *phstream, unsigned int flags){
-    LOG_INFO("cuStreamCreate %p",phstream);
     CUresult res = CUDA_OVERRIDE_CALL(cuda_library_entry,cuStreamCreate,phstream,flags);
     return res;
 }